arxiv-plugin/podcast_backend/requirements.txt
flask
flask-cors
PyPDF2
google-generativeai>=0.8.0
requests
python-dotenv
